Move assertObjectType to the one place it's used.
